fix(sim): GPS 经纬度改 f64，消除 M 场 NMEA 往返的额外量化（两场输入等价）
问题（已量化）：M 场把 GPS 走 NED -> float32 lat/lon -> NMEA -> 解析回 NED，
而 FlySimState.gps_lat/lon 是 f32。在真实坐标量级下 f32 的 ulp 折算为
  lat 31.2304  -> 0.212 m
  lon 121.4737 -> 0.726 m
  alt 4.0      -> ~0
而 SensorConfig::realistic() 的 GPS 位置噪声只有 0.5 m => 经度量化是噪声的 1.45 倍，
是**额外叠加**的噪声。H 场直传 NED PosSample，没有这一层 => 两场输入不等价。
且这是**仿真侧的人为缺陷**：真实 GPS 的 NMEA 精度远高于 f32 存储（ddmm.mmmmm
约 1.9mm），固件解析 NMEA 本身是该被测的真实路径，应保留。

修法（最小侵入，不动物理）：加一个 f64 取值口，只让经纬度走它。
- SensorModel::value64()：默认回落 `value() as f64`，其余通道零改动；
- FlySimSource 覆写 value64：lat/lon 直读 st.gps_lat/lon（f64）；
- DataSource::value64() + NMEA 生成器改用它；
- FlySimState.gps_lat/lon: f32 -> f64，并修 5 处写入点。
量化从此只剩 NMEA 文本本身（约 1.9mm，可忽略）。

注：examples/mpu_violation.rs 编译失败是**既有**问题（调用已改名的 m.run），
与本次改动无关，未在本提交中处理。

+++ src/env/scenario.rs
@@ -492,8 +492,8 @@ impl EnvScenario {
             }
             let dlat = gps_pos[0] / 111_320.0;
             let dlon = gps_pos[1] / (111_320.0 * self.lat0.to_radians().cos());
-            st.gps_lat = self.lat0 + dlat;
-            st.gps_lon = self.lon0 + dlon;
+            st.gps_lat = self.lat0 as f64 + dlat as f64;
+            st.gps_lon = self.lon0 as f64 + dlon as f64;
             st.gps_alt = self.alt_ref - gps_pos[2];
             let mut gvel = tr.vel;
             if let Some(nz) = self.perturb.noise {
